dashboard-back/requirements.txt
django==3.2.9
django-environ==0.8.1
pyjwt==2.3.0
requests==2.26.0
requests-oauthlib==1.3.0
django-cryptography==1.0